Impelements the structute od the MVC
