+++ src/main.java
@@ -1,5 +1,4 @@
-import models.gui.BattleSetupFrame;
-
+import view.BattleSetupFrame;
 import javax.swing.*;
 
 public static void main(String[] args) {
